data model fix

+++ src/bulle_planning_model/extractors/mengenlisten_extractor/mengenliste_entry.py
@@ -1,10 +1,10 @@
 from pydantic import BaseModel, Field
-from typing import Optional
+from typing import Optional, Union
 
 
 class MengenlisteEntry(BaseModel):
     """Represents a single article entry in a Mengenliste shift report."""
     article_name: str = Field(..., description="Product name")
     stock: Optional[int] = Field(None, description="Current stock quantity (Aktuelle Menge)")
-    leftover: Optional[int] = Field(None, description="Leftover quantity (Retoure Anzahl)")
+    leftover: Optional[float] = Field(None, description="Leftover quantity (Retoure Anzahl)")
     sold_out: Optional[str] = Field(None, description="Time when sold out (Ausverkauft um)")
